Medical_Assistant_langchain/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.2
supabase>=1.0.4
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.1
